+++ lib/jekyll/page.rb
@@ -143,9 +143,14 @@ module Jekyll
     #
     # Returns the destination file path String.
     def destination(dest)
+<<<<<<< HEAD
       path = site.in_dest_dir(dest, URL.unescape_path(url))
       path = File.join(path, "index") if url.end_with?("/")
       path <<  output_ext unless path.end_with?(output_ext)
+=======
+      path = Jekyll.sanitized_path(dest, url)
+      path = File.join(path, "index.html") if url =~ /\/$/
+>>>>>>> jekyll/v1-stable
       path
     end