+++ lib/jekyll/url.rb
@@ -70,8 +70,16 @@ module Jekyll
       end
     end
 
+<<<<<<< HEAD
     # Returns a sanitized String URL, stripping "../../" and multiples of "/",
     # as well as the beginning "/" so we can enforce and ensure it.
+=======
+    # Returns a sanitized String URL
+    def sanitize_url(in_url)
+
+      # Remove all double slashes
+      url = in_url.gsub(/\/\//, "/")
+>>>>>>> jekyll/v1-stable
 
     def sanitize_url(str)
       "/" + str.gsub(/\/{2,}/, "/").gsub(%r!\.+\/|\A/+!, "")
@@ -101,6 +109,7 @@ module Jekyll
       URI.escape(path, /[^a-zA-Z\d\-._~!$&'()*+,;=:@\/]/).encode('utf-8')
     end
 
+<<<<<<< HEAD
     # Unescapes a URL path segment
     #
     # path - The path to be unescaped.
@@ -113,6 +122,12 @@ module Jekyll
     # Returns the unescaped path.
     def self.unescape_path(path)
       URI.unescape(path.encode('utf-8'))
+=======
+      # Always add a leading slash
+      url.gsub!(/\A([^\/])/, '/\1')
+
+      url
+>>>>>>> jekyll/v1-stable
     end
   end
 end